Fix header folders hierarchy
