+++ WBooks/App.js
@@ -8,20 +8,18 @@
 
 import React from 'react';
 import {
-  SafeAreaView,
   StatusBar,
-  View,
 } from 'react-native';
 
 import BookList from './src/app/screens/Library/screens';
 import Header from './src/app/components/Header';
 
-const App =() => {
+function App() {
   return (
     <>
       <StatusBar barStyle='light-content' />
       <Header />
-        <BookList />
+      <BookList />
     </>
   );
 };
